Refactoring - moving classes around.
