+++ membership-attribute-service/app/services/aws/AwsS3Client.scala
@@ -1,4 +1,4 @@
-package aws
+package services.catalog.aws
 
 import com.amazonaws.services.s3.{AmazonS3, AmazonS3Client}
 import com.amazonaws.services.s3.model.{GetObjectRequest, S3ObjectInputStream}
