+++ membership-attribute-service/app/services/aws/package.scala
@@ -1,3 +1,5 @@
+package services.catalog
+
 import com.amazonaws.auth.{InstanceProfileCredentialsProvider, AWSCredentialsProviderChain}
 import com.amazonaws.auth.profile.ProfileCredentialsProvider
