+++ membership-attribute-service/app/services/payment/PaymentService.scala
@@ -1,11 +1,11 @@
-package models.subscription.services.api
+package services.payment
 
 import _root_.services.stripe.StripeService
-import _root_.services.zuora.soap.models.Queries.Account
-import models.PaymentDetails
 import models.subscription.Subscription.{AccountId, Id}
-import models.subscription.subsv2.{Subscription, SubscriptionPlan}
 import models.subscription._
+import models.subscription.subsv2.{Subscription, SubscriptionPlan}
+import models.PaymentDetails
+import _root_.services.zuora.soap.models.Queries.Account
 import scalaz.\/
 
 import scala.concurrent.Future
