+++ membership-attribute-service/app/services/payment/ZuoraPaymentService.scala
@@ -1,30 +1,31 @@
-package models.subscription.services
+package services.payment
 
-import monitoring.SafeLogger
-import monitoring.SafeLogger.Sanitizer
-import services.stripe.Stripe.Customer
-import services.stripe.{Stripe, StripeService}
+import _root_.services.zuora.soap.ZuoraSoapService
+import _root_.services.zuora.soap.models.Queries
+import _root_.services.zuora.soap.models.Queries.Account
+import _root_.services.zuora.soap.models.Queries.PaymentMethod._
 import models.PaymentDetails
-import services.zuora.soap.models.Queries
-import services.zuora.soap.models.Queries.Account
-import services.zuora.soap.models.Queries.PaymentMethod._
+import models.PaymentDetails.Payment
 import models.subscription.Subscription._
 import models.subscription.subsv2.SubscriptionPlan.Contributor
 import models.subscription.subsv2.{Subscription, SubscriptionPlan}
 import models.subscription.{BillingSchedule, Subscription => _, _}
+import monitoring.SafeLogger
+import monitoring.SafeLogger.Sanitizer
 import scalaz.std.option._
 import scalaz.std.scalaFuture._
 import scalaz.syntax.monad._
 import scalaz.syntax.std.option._
 import scalaz.{MonadTrans, OptionT, \/}
+import services.payment
+import _root_.services.stripe.{Stripe, StripeService}
+import services.stripe.Stripe.Customer
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Failure, Success, Try}
-import _root_.services.zuora.soap.ZuoraSoapService
-import models.PaymentDetails.Payment
 
-class PaymentService(zuoraService: ZuoraSoapService, planMap: Map[ProductRatePlanChargeId, Benefit])(implicit ec: ExecutionContext)
-    extends api.PaymentService {
+class ZuoraPaymentService(zuoraService: ZuoraSoapService, planMap: Map[ProductRatePlanChargeId, Benefit])(implicit ec: ExecutionContext)
+    extends payment.PaymentService {
 
   implicit val monadTrans = MonadTrans[OptionT] // it's the only one we use here, really
