+++ membership-attribute-service/app/services/stripe/BasicStripeServiceWithMetrics.scala
@@ -1,9 +1,8 @@
 package services.stripe
 
 import com.gu.i18n.Currency
-import services.stripe.Stripe
-import services.stripe.Stripe.{Charge, Customer, CustomersPaymentMethods, StripeObject}
 import monitoring.CreateMetrics
+import services.stripe.Stripe.{Charge, Customer, CustomersPaymentMethods, StripeObject}
 
 import scala.concurrent.{ExecutionContext, Future}
