+++ membership-attribute-service/app/wiring/AppLoader.scala
@@ -2,7 +2,6 @@ package wiring
 
 import actions.CommonActions
 import akka.actor.ActorSystem
-import models.subscription.subsv2.services.CatalogService
 import components.TouchpointBackends
 import configuration.{CreateTestUsernames, LogstashConfig, SentryConfig, Stage}
 import controllers._
@@ -17,6 +16,7 @@ import play.api.mvc.EssentialFilter
 import play.filters.cors.{CORSConfig, CORSFilter}
 import play.filters.csrf.CSRFComponents
 import router.Routes
+import services.catalog.CatalogService
 import services.salesforce.ContactRepository
 import services.stripe.BasicStripeService
 import services.subscription.SubscriptionService
